% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMLEfunction.R
\name{adjustTykohonov}
\alias{adjustTykohonov}
\title{Adjusts tykohonov parameter}
\usage{
adjustTykohonov(obsDiff, obsmean, mu, selected, firstDiff, secondDiff,
  tykohonvSlack, regularization_param)
}
\description{
Adjusts tykohonov parameter
}
