% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_data_generation.R
\name{find_clusters}
\alias{find_clusters}
\title{find clusters of adjascent selected coordinates}
\usage{
find_clusters(distances, selected_coordinates, magnitude_threshold,
  dist_threshold = 2, dist_method = "manhattan")
}
\arguments{
\item{selected_coordinates}{an integer vector, the coordinates of which
denotes which coordinates have crossed the threshold (have been selected)
such that will be obtained, for example from calling
\code{which(x > threshold)}.}
}
\description{
find clusters of adjascent selected coordinates
}
