% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsamplers.R
\name{mvtnormSelectionProbs}
\alias{mvtnormSelectionProbs}
\title{A function for computing the probability of positive/negative selections}
\usage{
mvtnormSelectionProbs(y, mu, sigma, threshold, selected = NULL)
}
\arguments{
\item{y}{the observed normal vector, or an initial sample point that
satisfies the selection event}

\item{mu}{the mean vector of the (untruncated) normal vector}

\item{sigma}{the covariance of the (untruncated) normal vector}

\item{threshold}{a vector of size length(y), the selection threshold}

\item{selected}{which coordinates were selected? if NULL function
will attempt to figure it out on its own but my fail if there are
both observations that are larger than \code{threshold} and observations
that are smaller than \code{-threshold}.}
}
\description{
This function uses the mvtnorm package to compute the probability of
positive and negative selection events of selective ROI problems.
The selection event is assumed to be, {selected > threshold} or
{selected < -threshold} (with coordinates which were not selected not
satisfying the selection event)
}
