% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMLEfunction.R
\name{roi_mle_control}
\alias{roi_mle_control}
\title{Generate a list of parameters controllong the stochastic gradient process of the \code{roiMLE} function}
\usage{
roi_mle_control(grad_iterations = 2100, step_size_coef = 0.5,
  step_rate = 0.55, samp_per_iter = 20, grad_delay = NULL,
  assume_convergence = NULL, impute_boundary = c("smooth", "neighbors",
  "none", "mean"))
}
\arguments{
\item{grad_iterations}{the number of stochastic gradient steps to take}

\item{step_size_coef}{step size coefficients for stochastic gradient step.
Best left unchanged.}

\item{step_rate}{the rate at which the stochastic gradient steps size should
decrease as a function of the number of steps already taken.}

\item{samp_per_iter}{the number of slice MH samples to take for computing the
stochastic gradient estimate in each stochastic gradient step}

\item{grad_delay}{the number of iterations to wait before starting to decrease
the stochastic gradient step size}

\item{assume_convergence}{after how many gradient steps should we assume
convergence? The final MLE estimate will be the average of the last
\code{grad_iterations - assume_convergence} estimates}

\item{impute_boundary}{the boundary imputation method to use. See
description for details}
}
\description{
Generate a list of parameters controllong the stochastic gradient process of the \code{roiMLE} function
}
