% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMLEfunction.R
\name{roi_sampling_control}
\alias{roi_sampling_control}
\title{Generate a list of parameters for sampling from the estimated conditional distribution}
\usage{
roi_sampling_control(samp_per_chain = 0, burnin = 1000, n_chains = 5)
}
\arguments{
\item{samp_per_chain}{the number of samples to take from the estimated
post-selection distribution in each MCMC chain. If 0, then no sampling
will be conducted (default).}

\item{burnin}{the burnin periods for each MCMC chain.}

\item{n_chains}{the number of MCMC chains to run. The MCMC chains use
different intializations. One chain will usually suffice, but it is
recommended to run several if there is strong depedence in the data
that may slow down the mixing of the MCMC chain.}
}
\description{
A function which generates a list of parameters for use in the roiMLE function,
for sampling from the estimated distribution. This can be used for constructing
confidence intervals or conducting hypothesis testing.
}
