% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMLEfunction.R
\name{roiMLE}
\alias{roiMLE}
\title{Compute the MLE for a selected region of interest}
\usage{
roiMLE(y, cov, threshold, coordinates = NULL, selected = NULL,
  mean_weights = NULL, projected = NULL, regularization_param = NULL,
  regularization_slack = 1, init = NULL, progress = FALSE,
  sampling_control = roi_sampling_control(),
  mle_control = roi_mle_control())
}
\arguments{
\item{y}{the observed noraml coordinates}

\item{cov}{the covariance of \code{y}}

\item{threshold}{the threshold used in the selection rule.
Must be either a scalar or a numeric vector of size \code{length(y)}}

\item{coordinates}{an optional matrix of the coordinates of the observed
vector. This is only relevant if \code{y} corresponds to a spatial
or temporal observation}

\item{selected}{an optional boolean vector, with \code{TRUE} coordinates
corresponding to coordinates of \code{y} that were selected.}

\item{mean_weights}{the weights to use for the contrast to be computed, if
not specified then equal weights will be given to all selected coordinates.
\code{mean_weights} do not have to sum to one. If specified, then the length
of the vector must either equal \code{sum(selected)} (the number of selected
coordinates), or \code{length(y)}. Either way, all coordinates which were not
selected must be given a weight of zero.}

\item{projected}{an optional fixed value that \code{mean(mu)} must equal.
Can be used to construct profile likelihood post-selection confidence
intervals.}

\item{regularization_param}{an optional penalty value for the tykohonov
regularizer. This is an (inferior) altenative to specifying a
\code{regularization_slack}.}

\item{regularization_slack}{the estimation routine uses first differences
Tykohonov regularization to estimate the mean of the selected region.
\code{regularization_slack} speficies the allowed deviation from the observed
first order differences. The description for details}

\item{init}{initial value for the mean estimate}

\item{progress}{whether to display a bar describing the progress
of the gradient algorithm.}

\item{sampling_control}{a list with control parameters for sampling
from the estimated distribution.}

\item{mle_contorl}{a list of parameters to be used when computed the
conditional MLE.}
}
\description{
Computes the conditional mle for a region selected based on
the selection rule \code{y[selected] > threshold} or
\code{y[selected] < -threshold}, and the coordinates which were not
selected must violate the selection rule.
}
