% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMLEfunction.R
\name{roiMLE}
\alias{roiMLE}
\title{Compute the MLE for a selected region of interest}
\usage{
roiMLE(y, cov, threshold, coordinates = NULL, selected = NULL,
  projected = NULL, tykohonovParam = NULL, tykohonovSlack = 1,
  stepSizeCoef = 0.15, stepRate = 0.65, sampPerIter = 40, delay = 100,
  maxiter = 2000, assumeConvergence = 1500, nsamp = 0, init = NULL,
  progress = FALSE, imputeBoundary = c("smooth", "neighbors", "none",
  "mean"))
}
\arguments{
\item{y}{the observed noraml coordinates}

\item{cov}{the covariance of \code{y}}

\item{threshold}{the threshold used in the selection rule.
Must be either a scalar or a numeric vector of size \code{length(y)}}

\item{coordinates}{an optional matrix of the coordinates of the observed
vector. This is only relevant if \code{y} corresponds to a spatial
or temporal observation}

\item{selected}{an optional boolean vector, with \code{TRUE} coordinates
corresponding to coordinates of \code{y} that were selected.}

\item{projected}{an optional fixed value that \code{mean(mu)} must equal.
Can be used to construct profile likelihood post-selection confidence
intervals.}

\item{tykohonovParam}{an optional penalty value for the tykohonov
regularizer. This is an (inferior) altenative to specifying a
\code{tykohonovSlack}.}

\item{tykohonovSlack}{the estimation routine uses first differences
Tykohonov regularization to estimate the mean of the selected region.
\code{tykohonovSlack} speficies the allowed deviation from the observed
first order differences. The description for details}

\item{stepSizeCoef}{step size coefficients for stochastic gradient step.
Best left unchanged.}

\item{stepRate}{the rate at which the stochastic gradient steps size should
decrease as a function of the number of steps already taken.}

\item{sampPerIter}{the number of slice MH samples to take for computing the
stochastic gradient estimate in each stochastic gradient step}

\item{delay}{the number of iterations to wait before starting to decrease
the stochastic gradient step size}

\item{maxiter}{the number of stochastic gradient steps to take}

\item{assumeConvergence}{after how many gradient steps should we assume
convergence? The final MLE estimate will be the average of the last
\code{maxiter - assumeConvergence} estimates}

\item{nsamp}{the number of samples to take from the estimated
post-selection distribution}

\item{init}{initial value for the mean estimate}

\item{imputeBoundary}{the boundary imputation method to use. See
description for details}
}
\description{
Computes the conditional mle for a region selected based on
the selection rule \code{y[selected] > threshold} or
\code{y[selected] < -threshold}, and the coordinates which were not
selected must violate the selection rule.
}
