% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMLEfunction.R
\name{roiMLE}
\alias{roiMLE}
\title{Compute the MLE for a selected region of interest}
\usage{
roiMLE(y, cov, threshold, coordinates = NULL, selected = NULL,
  projected = NULL, regularization_param = NULL, regularization_slack = 1,
  init = NULL, progress = FALSE, control = mle_control())
}
\arguments{
\item{y}{the observed noraml coordinates}

\item{cov}{the covariance of \code{y}}

\item{threshold}{the threshold used in the selection rule.
Must be either a scalar or a numeric vector of size \code{length(y)}}

\item{coordinates}{an optional matrix of the coordinates of the observed
vector. This is only relevant if \code{y} corresponds to a spatial
or temporal observation}

\item{selected}{an optional boolean vector, with \code{TRUE} coordinates
corresponding to coordinates of \code{y} that were selected.}

\item{projected}{an optional fixed value that \code{mean(mu)} must equal.
Can be used to construct profile likelihood post-selection confidence
intervals.}

\item{regularization_param}{an optional penalty value for the tykohonov
regularizer. This is an (inferior) altenative to specifying a
\code{regularization_slack}.}

\item{regularization_slack}{the estimation routine uses first differences
Tykohonov regularization to estimate the mean of the selected region.
\code{regularization_slack} speficies the allowed deviation from the observed
first order differences. The description for details}

\item{init}{initial value for the mean estimate}

\item{step_size_coef}{step size coefficients for stochastic gradient step.
Best left unchanged.}

\item{step_rate}{the rate at which the stochastic gradient steps size should
decrease as a function of the number of steps already taken.}

\item{samp_per_iter}{the number of slice MH samples to take for computing the
stochastic gradient estimate in each stochastic gradient step}

\item{grad_delay}{the number of iterations to wait before starting to decrease
the stochastic gradient step size}

\item{grad_iterations}{the number of stochastic gradient steps to take}

\item{assume_convergence}{after how many gradient steps should we assume
convergence? The final MLE estimate will be the average of the last
\code{grad_iterations - assume_convergence} estimates}

\item{nsamp}{the number of samples to take from the estimated
post-selection distribution}

\item{impute_boundary}{the boundary imputation method to use. See
description for details}
}
\description{
Computes the conditional mle for a region selected based on
the selection rule \code{y[selected] > threshold} or
\code{y[selected] < -threshold}, and the coordinates which were not
selected must violate the selection rule.
}
